16th March Update
- Gen 3 WIP
- Ninjask / Shedinja evolution
- Added some alerts when you evolve your Pokemon
- Start to setup the Security Update, it's a placeholder.

+++ stats.js
@@ -113,12 +113,14 @@ function exportSave() {
   const blob = new Blob([encryptedSaveData], { type: 'text/plain' });
   const url = URL.createObjectURL(blob);
 
-  const link = document.createElement('a');
-  link.href = url;
-  link.download = 'PHC-save.txt';
-  link.click();
-  
-  localStorage.setItem('lastExportTimestamp', currentTimestamp.toString());
+  if (confirm('Are you sure you want to download the file? The cooldown will be triggered!')) {
+    const link = document.createElement('a');
+    link.href = url;
+    link.download = 'PHC-save.txt';
+    link.click();
+    
+    localStorage.setItem('lastExportTimestamp', currentTimestamp.toString());
+  }
 }
 
 function importSave() {
@@ -129,7 +131,7 @@ function importSave() {
     alert('You can only import your data once every 8 hours.');
     return;
   }
-  
+
   const input = document.createElement('input');
   input.type = 'file';
   input.accept = 'text/plain';
@@ -151,9 +153,13 @@ function importSave() {
     };
     reader.readAsText(file);
   };
-  input.click();
   
-  localStorage.setItem('lastImportTimestamp', currentTimestamp.toString());
+  // Confirm import with user
+  const confirmImport = confirm('Are you sure you want to import save data? This will overwrite your current save data.');
+  if (confirmImport) {
+    input.click();
+    localStorage.setItem('lastImportTimestamp', currentTimestamp.toString());
+  }
 }
 
 /*function reloadLocalStorage() {
@@ -168,7 +174,7 @@ function importSave() {
 
 reloadLocalStorage();*/
 
-function _0x15b3(_0x4df2af,_0x11c365){const _0x2487f5=_0x2487();return _0x15b3=function(_0x15b386,_0x45b686){_0x15b386=_0x15b386-0x69;let _0x155440=_0x2487f5[_0x15b386];return _0x155440;},_0x15b3(_0x4df2af,_0x11c365);}const _0x2cac0d=_0x15b3;(function(_0x3d0878,_0x531827){const _0x44ffe6=_0x15b3,_0x5826bf=_0x3d0878();while(!![]){try{const _0xc0b1af=-parseInt(_0x44ffe6(0x6a))/0x1*(parseInt(_0x44ffe6(0x73))/0x2)+-parseInt(_0x44ffe6(0x69))/0x3+-parseInt(_0x44ffe6(0x6d))/0x4*(-parseInt(_0x44ffe6(0x71))/0x5)+parseInt(_0x44ffe6(0x72))/0x6+parseInt(_0x44ffe6(0x74))/0x7+parseInt(_0x44ffe6(0x70))/0x8*(-parseInt(_0x44ffe6(0x6b))/0x9)+-parseInt(_0x44ffe6(0x6e))/0xa*(-parseInt(_0x44ffe6(0x6c))/0xb);if(_0xc0b1af===_0x531827)break;else _0x5826bf['push'](_0x5826bf['shift']());}catch(_0x3fb17b){_0x5826bf['push'](_0x5826bf['shift']());}}}(_0x2487,0xd37c1));function _0x2487(){const _0x59d379=['5441664wSXSGM','228318YQyBTP','2689834ivslgX','3313491nrBEAH','1hfTQVU','9NwfEfJ','337315NQLEfj','677188ODzUWi','470RIzRZX','AreYouKeepingUpWithTheCommodore','11952392quXIHZ','25mkfsnW'];_0x2487=function(){return _0x59d379;};return _0x2487();}const secretKey=_0x2cac0d(0x6f);
+const _0xcbfe40=_0x1376;function _0x3fce(){const _0x33f211=['2212884wSXdsq','669837tViMfR','40BoBHlH','CauseTheCommodoreIsKeepingUpWithYou','3591424UEbWgn','303704YBQXvY','1247106exeMki','278156CMPkJf','1482366rqIDvW'];_0x3fce=function(){return _0x33f211;};return _0x3fce();}(function(_0x558281,_0x10dc23){const _0x57f7bd=_0x1376,_0x5aa1b4=_0x558281();while(!![]){try{const _0x436f07=-parseInt(_0x57f7bd(0x1dd))/0x1+parseInt(_0x57f7bd(0x1d5))/0x2+-parseInt(_0x57f7bd(0x1d6))/0x3+parseInt(_0x57f7bd(0x1db))/0x4*(parseInt(_0x57f7bd(0x1d8))/0x5)+-parseInt(_0x57f7bd(0x1dc))/0x6+-parseInt(_0x57f7bd(0x1d7))/0x7+parseInt(_0x57f7bd(0x1da))/0x8;if(_0x436f07===_0x10dc23)break;else _0x5aa1b4['push'](_0x5aa1b4['shift']());}catch(_0x3a4a2e){_0x5aa1b4['push'](_0x5aa1b4['shift']());}}}(_0x3fce,0x74bf1));function _0x1376(_0x314f04,_0x449961){const _0x3fcecc=_0x3fce();return _0x1376=function(_0x13766e,_0x55a080){_0x13766e=_0x13766e-0x1d5;let _0x2d1598=_0x3fcecc[_0x13766e];return _0x2d1598;},_0x1376(_0x314f04,_0x449961);}const secretKey=_0xcbfe40(0x1d9);
 
 /*function updatePokemonSprites() {
   [team, storage].forEach((pokemonArray) => {
